requirements.txt
click
python-vlc
readchar
